Roll back checking whether scheduler's threads are consistent.
Since this VM image does not have SMP, all threads share the
same reference to the global runqueue. Accessing this global
runqueue caused issues (even though we check whether the status
bit of the specific runqueue is set). The linked lists of different
runqueues sometimes point to invalid memory despite having
checked the status bit.

Iterating through the sleepqueue is not useful as well,
because the reference of the curthread to its sleepqueue
is invalid (as the curthread is running and not sleeping).